+++ backend/src/routes/user/login/route.ts
@@ -1,26 +1,11 @@
-import { RequestHandler } from "express";
-
 import passport from "passport";
+
+import { isNotLoggedIn } from "@middlewares/user";
 
 import { userRouter } from "..";
 
 import { ReqDTO, ResDTO } from "./types";
 
-export const isLoggedIn: RequestHandler<object> = (req, res, next) => {
-  if (req.isAuthenticated()) {
-    next();
-  } else {
-    res.status(403).send("로그인이 필요합니다.");
-  }
-};
-
-export const isNotLoggedIn: RequestHandler<object> = (req, res, next) => {
-  if (!req.isAuthenticated()) {
-    next();
-  } else {
-    res.redirect("/");
-  }
-};
 userRouter.post<object, ResDTO, ReqDTO>(
   "/login",
   isNotLoggedIn,
