+++ backend/tsconfig.paths.json
@@ -5,7 +5,8 @@
       "@services/*": ["./src/services/*"],
       "@models/*": ["./src/models/*"],
       "@routes": ["./src/routes"],
-      "@controllers/*": ["./src/controllers/*"]
+      "@controllers/*": ["./src/controllers/*"],
+      "@middlewares/*": ["./src/middlewares/*"]
     }
   }
 }
